+++ src/encoder_sse_base.h
@@ -103,7 +103,7 @@ template<enum YEncDecIsaLevel use_isa>
 static size_t do_encode_sse(int line_size, int* colOffset, const unsigned char* src, unsigned char* dest, size_t len) {
 	unsigned char* es = (unsigned char*)src + len;
 	unsigned char *p = dest; // destination pointer
-	long i = -len; // input position
+	long i = -(long)len; // input position
 	unsigned char c, escaped; // input character; escaped input character
 	int col = *colOffset;
